refactor: split client and page
